Add github actions to replace travis
Currently points to my forks of AR and SMW, but I hope to get those changes merged soon.

Fixes #17
