+++ .editorconfig
@@ -1,0 +1,11 @@
+root = true
+
+[*]
+indent_style = tab
+tab_width = 4
+end_of_line = lf
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+[*.y{a,}ml]
+indent_style = space
