--- .scrutinizer.yml
@@ -1,16 +1,0 @@
-inherit: true
-
-tools:
-    php_code_coverage: false
-    external_code_coverage: true
-    php_code_sniffer: true
-    php_cpd: true
-    php_cs_fixer: true
-    php_loc: true
-    php_mess_detector:
-        config:
-            controversial_rules: { superglobals: false }
-    php_cpd: true
-    php_pdepend: true
-    php_analyzer: true
-    sensiolabs_security_checker: true
