--- .travis.yml
@@ -1,39 +1,0 @@
-# http://docs.travis-ci.com/user/workers/container-based-infrastructure/
-sudo: false
-
-language: php
-
-# Travis support wrote "Ubuntu 14.04 reached EOL on April 30th, 2019,
-# ... we've been slowly rolling out changes to make Xenial builds ..."
-# https://blog.travis-ci.com/2019-04-15-xenial-default-build-environment
-# https://docs.travis-ci.com/user/reference/xenial/#services-disabled-by-default.
-services:
-  - mysql
-
-matrix:
-  fast_finish: true
-  include:
-    - env: DB=mysql; MW=REL1_35; PHPUNIT=8.5.*; APPROVED_REVS=1.0
-      php: 7.4
-    - env: DB=mysql; MW=REL1_35; TYPE=coverage; PHPUNIT=5.7.*; APPROVED_REVS=1.0
-      php: 7.3
-    - env: DB=mysql; MW=REL1_35; PHPUNIT=8.5.*; APPROVED_REVS=1.0
-      php: 7.2
-  allow_failures:
-    # Due to APPROVED_REVS failing on MW master
-    - env: DB=mysql; MW=master; PHPUNIT=6.5.*; APPROVED_REVS=1.0
-
-install:
-  - composer self-update
-  - bash ./tests/travis/install-mediawiki.sh
-  - bash ./tests/travis/install-semantic-approved-revs.sh
-
-script:
-  - bash ./tests/travis/run-tests.sh
-
-after_success:
-  - bash ./tests/travis/upload-coverage-report.sh
-
-cache:
-  directories:
-    - $HOME/.composer/cache
