+++ tests/phpunit/Unit/PropertyAnnotators/ApprovedByPropertyAnnotatorTest.php
@@ -18,7 +18,7 @@ class ApprovedByPropertyAnnotatorTest extends \PHPUnit_Framework_TestCase {
 
 	private $databaseLogReader;
 
-	protected function setUp() {
+	protected function setUp(): void {
 		parent::setUp();
 
 		$this->databaseLogReader = $this->getMockBuilder( '\SMW\ApprovedRevs\DatabaseLogReader' )
