+++ tests/phpunit/Unit/ServicesFactoryTest.php
@@ -61,7 +61,7 @@ class ServicesFactoryTest extends \PHPUnit_Framework_TestCase {
 		$appFactory->setConnection( $connection );
 
 		$dbLogReader = $appFactory->newDatabaseLogReader( null );
-		$dbLogReader->getStatusOfLogEntry();
+		$this->assertNull( $dbLogReader->getStatusOfLogEntry() );
 	}
 
 	/**
